@@ -4,11 +4,11 @@
   "description": "Data Tiles is a small website that shows data.",
   "main": "server.ts",
   "scripts": {
-    "build": "next build",
-    "start": "cross-env NODE_ENV=production ts-node server.ts",
+    "build": "next build && tsc -p tsconfig.server.json",
+    "start": "cross-env NODE_ENV=production ts-node -P tsconfig.server.json .next/server/index.js",
     "test": "cross-env NODE_ENV=test jest",
     "coverage": "cross-env NODE_ENV=test jest --coverage",
-    "dev": "nodemon server.ts --exec babel-node --extensions '.ts'"
+    "dev": "nodemon server/index.ts"
   },
   "keywords": [
     "data",
@@ -18,53 +18,54 @@
   "author": "lovethebomb",
   "license": "ISC",
   "dependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/node": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@types/enzyme": "^3.1.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/node": "^7.2.2",
+    "@babel/preset-env": "^7.3.1",
+    "@types/enzyme": "^3.1.17",
     "@types/jest": "^24.0.0",
     "@types/next": "^8.0.0",
-    "@types/react": "^16.3.17",
-    "@types/react-dom": "^16.0.6",
-    "@types/react-transition-group": "^2.0.11",
-    "@types/styled-jsx": "^2.2.3",
+    "@types/react": "^16.8.2",
+    "@types/react-dom": "^16.8.0",
+    "@types/react-transition-group": "^2.0.15",
+    "@types/styled-jsx": "^2.2.7",
     "@zeit/next-typescript": "^1.1.1",
-    "apicache": "^1.2.1",
-    "babel-loader": "^8.0.0",
+    "apicache": "^1.2.3",
+    "babel-loader": "^8.0.5",
     "cross-env": "^5.2.0",
     "css-ease": "0.0.1",
-    "date-fns": "^2.0.0-alpha.10",
+    "date-fns": "^2.0.0-alpha.27",
     "del": "^3.0.0",
     "del-cli": "^1.1.0",
-    "dotenv": "^6.0.0",
+    "dotenv": "^6.2.0",
     "enzyme": "^3.6.0",
-    "enzyme-adapter-react-16": "^1.4.0",
-    "express": "^4.16.3",
+    "enzyme-adapter-react-16": "^1.9.1",
+    "express": "^4.16.4",
     "isomorphic-fetch": "^2.2.1",
     "isomorphic-unfetch": "^3.0.0",
-    "jest": "^23.1.0",
-    "morgan": "^1.9.0",
+    "jest": "^23.6.0",
+    "morgan": "^1.9.1",
     "next": "^7.0.3",
-    "npm": "^6.4.1",
+    "npm": "^6.7.0",
     "parse-prometheus-text-format": "^1.0.5",
-    "prom-client": "^11.0.0",
+    "prom-client": "^11.2.1",
     "quake-champions-api": "^1.0.5",
-    "react": "^16.4.1",
+    "react": "^16.8.1",
     "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^16.4.1",
-    "react-test-renderer": "^16.4.1",
-    "react-transition-group": "^2.3.1",
-    "tslint": "^5.10.0",
-    "tslint-config-prettier": "^1.13.0",
+    "react-dom": "^16.8.1",
+    "react-test-renderer": "^16.8.1",
+    "react-transition-group": "^2.5.3",
+    "tslint": "^5.12.1",
+    "tslint-config-prettier": "^1.18.0",
     "tslint-react": "^3.6.0",
     "typescript": "^2.9.2",
-    "typescript-babel-jest": "^1.0.5"
+    "typescript-babel-jest": "^1.0.6"
   },
   "devDependencies": {
-    "eslint": "^5.0.0",
+    "@types/apicache": "^1.2.0",
+    "eslint": "^5.13.0",
     "eslint-config-unsplash": "^0.2.2",
-    "nodemon": "^1.17.5",
+    "nodemon": "^1.18.10",
     "ts-jest": "^23.10.5",
     "ts-node": "^7.0.0"
   },
